+++ tasks/mpi/varfolomeev_g_transfer_from_one_to_all_scatter_custom/func_tests/main.cpp
@@ -5,9 +5,30 @@
 #include <boost/mpi/environment.hpp>
 #include <vector>
 
-#include "mpi/varfolomeev_g_transfer_from_one_to_all_scatter/include/ops_mpi.hpp"
+#include "mpi/varfolomeev_g_transfer_from_one_to_all_scatter_custom/include/ops_mpi.hpp"
 
-TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_pipeline) {
+TEST(varfolomeev_g_transfer_from_one_to_all_custom_scatter_mpi, Test_CustomScatter_pipeline) {
+  boost::mpi::communicator world;
+  std::vector<int> global_vec;            // working vector
+  std::vector<int32_t> global_res(1, 0);  // result vector
+  std::shared_ptr<ppc::core::TaskData> taskDataPar = std::make_shared<ppc::core::TaskData>();
+  if (world.rank() == 0) {
+    const int count_size_vector = 0;
+    global_vec = varfolomeev_g_transfer_from_one_to_all_scatter_mpi::getRandomVector(count_size_vector, -100, 100);
+    taskDataPar->inputs.emplace_back(reinterpret_cast<uint8_t*>(global_vec.data()));
+    taskDataPar->inputs_count.emplace_back(global_vec.size());
+    taskDataPar->outputs.emplace_back(reinterpret_cast<uint8_t*>(global_res.data()));
+    taskDataPar->outputs_count.emplace_back(global_res.size());
+  }
+  varfolomeev_g_transfer_from_one_to_all_custom_scatter_mpi::MyScatterTestMPITaskParallel MyScatterTestMPITaskParallel(
+      taskDataPar, "+");
+  ASSERT_EQ(MyScatterTestMPITaskParallel.validation(), true);
+  ASSERT_EQ(MyScatterTestMPITaskParallel.pre_processing(), true);
+  ASSERT_EQ(MyScatterTestMPITaskParallel.run(), true);
+  ASSERT_EQ(MyScatterTestMPITaskParallel.post_processing(), true);
+}
+
+TEST(varfolomeev_g_transfer_from_one_to_all_custom_scatter_mpi, Test_CustomScatter_Validation_Wrong_Operation) {
   boost::mpi::communicator world;
   std::vector<int> global_vec;            // working vector
   std::vector<int32_t> global_res(1, 0);  // result vector
@@ -22,20 +43,20 @@ TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_pipeline) {
     taskDataPar->outputs_count.emplace_back(global_res.size());
   }
 
-  varfolomeev_g_transfer_from_one_to_all_scatter_mpi::TestMPITaskParallel testMpiTaskParallel(taskDataPar, "+");
+  varfolomeev_g_transfer_from_one_to_all_custom_scatter_mpi::MyScatterTestMPITaskParallel MyScatterTestMPITaskParallel(
+      taskDataPar, "min");
 
-  ASSERT_EQ(testMpiTaskParallel.validation(), true);
-  ASSERT_EQ(testMpiTaskParallel.pre_processing(), true);
-  ASSERT_EQ(testMpiTaskParallel.run(), true);
-  ASSERT_EQ(testMpiTaskParallel.post_processing(), true);
+  ASSERT_EQ(MyScatterTestMPITaskParallel.validation(), false);
+  ASSERT_EQ(MyScatterTestMPITaskParallel.pre_processing(), true);
+  ASSERT_EQ(MyScatterTestMPITaskParallel.run(), true);
+  ASSERT_EQ(MyScatterTestMPITaskParallel.post_processing(), true);
 }
 
-TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Validation_Wrong_Operation) {
+TEST(varfolomeev_g_transfer_from_one_to_all_custom_scatter_mpi, Test_CustomScatter_Empty_vec) {
   boost::mpi::communicator world;
   std::vector<int> global_vec;            // working vector
   std::vector<int32_t> global_res(1, 0);  // result vector
   std::shared_ptr<ppc::core::TaskData> taskDataPar = std::make_shared<ppc::core::TaskData>();
-
   if (world.rank() == 0) {
     const int count_size_vector = 0;
     global_vec = varfolomeev_g_transfer_from_one_to_all_scatter_mpi::getRandomVector(count_size_vector, -100, 100);
@@ -44,51 +65,12 @@ TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Validation_Wrong_O
     taskDataPar->outputs.emplace_back(reinterpret_cast<uint8_t*>(global_res.data()));
     taskDataPar->outputs_count.emplace_back(global_res.size());
   }
-
-  varfolomeev_g_transfer_from_one_to_all_scatter_mpi::TestMPITaskParallel testMpiTaskParallel(taskDataPar, "min");
-
-  ASSERT_EQ(testMpiTaskParallel.validation(), false);
-  ASSERT_EQ(testMpiTaskParallel.pre_processing(), true);
-  ASSERT_EQ(testMpiTaskParallel.run(), true);
-  ASSERT_EQ(testMpiTaskParallel.post_processing(), true);
-}
-
-TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Generator) {
-  boost::mpi::communicator world;
-  std::vector<int> global_vec;  // working vector
-  int a = -100;
-  int b = 100;
-  int count_size_vector = 500;
-  global_vec = varfolomeev_g_transfer_from_one_to_all_scatter_mpi::getRandomVector(count_size_vector, a, b);
-  int max = *std::max_element(global_vec.begin(), global_vec.end());
-  int min = *std::min_element(global_vec.begin(), global_vec.end());
-
-  ASSERT_LE(max, b);
-  ASSERT_GE(min, a);
-  ASSERT_EQ(count_size_vector, (int)global_vec.size());
-}
-
-TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Empty_vec) {
-  boost::mpi::communicator world;
-  std::vector<int> global_vec;            // working vector
-  std::vector<int32_t> global_res(1, 0);  // result vector
-  std::shared_ptr<ppc::core::TaskData> taskDataPar = std::make_shared<ppc::core::TaskData>();
-
-  if (world.rank() == 0) {
-    const int count_size_vector = 0;
-    global_vec = varfolomeev_g_transfer_from_one_to_all_scatter_mpi::getRandomVector(count_size_vector, -100, 100);
-    taskDataPar->inputs.emplace_back(reinterpret_cast<uint8_t*>(global_vec.data()));
-    taskDataPar->inputs_count.emplace_back(global_vec.size());
-    taskDataPar->outputs.emplace_back(reinterpret_cast<uint8_t*>(global_res.data()));
-    taskDataPar->outputs_count.emplace_back(global_res.size());
-  }
-
-  varfolomeev_g_transfer_from_one_to_all_scatter_mpi::TestMPITaskParallel testMpiTaskParallel(taskDataPar, "+");
-  ASSERT_EQ(testMpiTaskParallel.validation(), true);
-  testMpiTaskParallel.pre_processing();
-  testMpiTaskParallel.run();
-  testMpiTaskParallel.post_processing();
-
+  varfolomeev_g_transfer_from_one_to_all_custom_scatter_mpi::MyScatterTestMPITaskParallel MyScatterTestMPITaskParallel(
+      taskDataPar, "+");
+  ASSERT_EQ(MyScatterTestMPITaskParallel.validation(), true);
+  MyScatterTestMPITaskParallel.pre_processing();
+  MyScatterTestMPITaskParallel.run();
+  MyScatterTestMPITaskParallel.post_processing();
   if (world.rank() == 0) {
     std::vector<int32_t> check_vec(1, 0);  // vector to check algorithm by seq. realisation
     std::shared_ptr<ppc::core::TaskData> taskDataSeq = std::make_shared<ppc::core::TaskData>();
@@ -96,26 +78,23 @@ TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Empty_vec) {
     taskDataSeq->inputs_count.emplace_back(global_vec.size());
     taskDataSeq->outputs.emplace_back(reinterpret_cast<uint8_t*>(check_vec.data()));
     taskDataSeq->outputs_count.emplace_back(check_vec.size());
-
     varfolomeev_g_transfer_from_one_to_all_scatter_mpi::TestMPITaskSequential testMpiTaskSequential(taskDataSeq, "+");
     ASSERT_EQ(testMpiTaskSequential.validation(), true);
     testMpiTaskSequential.pre_processing();
     testMpiTaskSequential.run();
     testMpiTaskSequential.post_processing();
-
     ASSERT_EQ(check_vec[0], global_res[0]);
   }
 }
 
-TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Sum_Zero_vec) {
+TEST(varfolomeev_g_transfer_from_one_to_all_custom_scatter_mpi, Test_CustomScatter_Sum_Zero_vec) {
   boost::mpi::communicator world;
   std::vector<int> global_vec;            // working vector
   std::vector<int32_t> global_res(1, 0);  // result vector
-
   std::shared_ptr<ppc::core::TaskData> taskDataPar = std::make_shared<ppc::core::TaskData>();
 
   if (world.rank() == 0) {
-    const int count_size_vector = 130;
+    const int count_size_vector = 555;
     global_vec = varfolomeev_g_transfer_from_one_to_all_scatter_mpi::getRandomVector(count_size_vector, 0, 0);
     taskDataPar->inputs.emplace_back(reinterpret_cast<uint8_t*>(global_vec.data()));
     taskDataPar->inputs_count.emplace_back(global_vec.size());
@@ -123,15 +102,14 @@ TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Sum_Zero_vec) {
     taskDataPar->outputs_count.emplace_back(global_res.size());
   }
 
-  varfolomeev_g_transfer_from_one_to_all_scatter_mpi::TestMPITaskParallel testMpiTaskParallel(taskDataPar, "+");
-  ASSERT_EQ(testMpiTaskParallel.validation(), true);
-  testMpiTaskParallel.pre_processing();
-  testMpiTaskParallel.run();
-  testMpiTaskParallel.post_processing();
-
+  varfolomeev_g_transfer_from_one_to_all_custom_scatter_mpi::MyScatterTestMPITaskParallel MyScatterTestMPITaskParallel(
+      taskDataPar, "+");
+  ASSERT_EQ(MyScatterTestMPITaskParallel.validation(), true);
+  MyScatterTestMPITaskParallel.pre_processing();
+  MyScatterTestMPITaskParallel.run();
+  MyScatterTestMPITaskParallel.post_processing();
   if (world.rank() == 0) {
     std::vector<int32_t> check_vec(1, 0);  // vector to check algorithm by seq. realisation
-
     std::shared_ptr<ppc::core::TaskData> taskDataSeq = std::make_shared<ppc::core::TaskData>();
     taskDataSeq->inputs.emplace_back(reinterpret_cast<uint8_t*>(global_vec.data()));
     taskDataSeq->inputs_count.emplace_back(global_vec.size());
@@ -148,58 +126,14 @@ TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Sum_Zero_vec) {
   }
 }
 
-TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Diff_Zero_vec) {
-  for (int i = 0; i < 10; i++) {
-    boost::mpi::communicator world;
-    std::vector<int> global_vec;            // working vector
-    std::vector<int32_t> global_res(1, 0);  // result vector
-
-    std::shared_ptr<ppc::core::TaskData> taskDataPar = std::make_shared<ppc::core::TaskData>();
-
-    if (world.rank() == 0) {
-      const int count_size_vector = 130;
-      global_vec = varfolomeev_g_transfer_from_one_to_all_scatter_mpi::getRandomVector(count_size_vector, 0, 0);
-      taskDataPar->inputs.emplace_back(reinterpret_cast<uint8_t*>(global_vec.data()));
-      taskDataPar->inputs_count.emplace_back(global_vec.size());
-      taskDataPar->outputs.emplace_back(reinterpret_cast<uint8_t*>(global_res.data()));
-      taskDataPar->outputs_count.emplace_back(global_res.size());
-    }
-
-    varfolomeev_g_transfer_from_one_to_all_scatter_mpi::TestMPITaskParallel testMpiTaskParallel(taskDataPar, "-");
-    ASSERT_EQ(testMpiTaskParallel.validation(), true);
-    testMpiTaskParallel.pre_processing();
-    testMpiTaskParallel.run();
-    testMpiTaskParallel.post_processing();
-
-    if (world.rank() == 0) {
-      std::vector<int32_t> check_vec(1, 0);  // vector to check algorithm by seq. realisation
-
-      std::shared_ptr<ppc::core::TaskData> taskDataSeq = std::make_shared<ppc::core::TaskData>();
-      taskDataSeq->inputs.emplace_back(reinterpret_cast<uint8_t*>(global_vec.data()));
-      taskDataSeq->inputs_count.emplace_back(global_vec.size());
-      taskDataSeq->outputs.emplace_back(reinterpret_cast<uint8_t*>(check_vec.data()));
-      taskDataSeq->outputs_count.emplace_back(check_vec.size());
-
-      varfolomeev_g_transfer_from_one_to_all_scatter_mpi::TestMPITaskSequential testMpiTaskSequential(taskDataSeq, "-");
-      ASSERT_EQ(testMpiTaskSequential.validation(), true);
-      testMpiTaskSequential.pre_processing();
-      testMpiTaskSequential.run();
-      testMpiTaskSequential.post_processing();
-
-      ASSERT_EQ(check_vec[0], global_res[0]);
-    }
-  }
-}
-
-TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Max_Zero_vec) {
+TEST(varfolomeev_g_transfer_from_one_to_all_custom_scatter_mpi, Test_CustomScatter_Diff_Zero_vec) {
   boost::mpi::communicator world;
   std::vector<int> global_vec;            // working vector
   std::vector<int32_t> global_res(1, 0);  // result vector
-
   std::shared_ptr<ppc::core::TaskData> taskDataPar = std::make_shared<ppc::core::TaskData>();
 
   if (world.rank() == 0) {
-    const int count_size_vector = 130;
+    const int count_size_vector = 50;
     global_vec = varfolomeev_g_transfer_from_one_to_all_scatter_mpi::getRandomVector(count_size_vector, 0, 0);
     taskDataPar->inputs.emplace_back(reinterpret_cast<uint8_t*>(global_vec.data()));
     taskDataPar->inputs_count.emplace_back(global_vec.size());
@@ -207,15 +141,53 @@ TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Max_Zero_vec) {
     taskDataPar->outputs_count.emplace_back(global_res.size());
   }
 
-  varfolomeev_g_transfer_from_one_to_all_scatter_mpi::TestMPITaskParallel testMpiTaskParallel(taskDataPar, "max");
-  ASSERT_EQ(testMpiTaskParallel.validation(), true);
-  testMpiTaskParallel.pre_processing();
-  testMpiTaskParallel.run();
-  testMpiTaskParallel.post_processing();
-
+  varfolomeev_g_transfer_from_one_to_all_custom_scatter_mpi::MyScatterTestMPITaskParallel MyScatterTestMPITaskParallel(
+      taskDataPar, "-");
+  ASSERT_EQ(MyScatterTestMPITaskParallel.validation(), true);
+  MyScatterTestMPITaskParallel.pre_processing();
+  MyScatterTestMPITaskParallel.run();
+  MyScatterTestMPITaskParallel.post_processing();
   if (world.rank() == 0) {
     std::vector<int32_t> check_vec(1, 0);  // vector to check algorithm by seq. realisation
+    std::shared_ptr<ppc::core::TaskData> taskDataSeq = std::make_shared<ppc::core::TaskData>();
+    taskDataSeq->inputs.emplace_back(reinterpret_cast<uint8_t*>(global_vec.data()));
+    taskDataSeq->inputs_count.emplace_back(global_vec.size());
+    taskDataSeq->outputs.emplace_back(reinterpret_cast<uint8_t*>(check_vec.data()));
+    taskDataSeq->outputs_count.emplace_back(check_vec.size());
 
+    varfolomeev_g_transfer_from_one_to_all_scatter_mpi::TestMPITaskSequential testMpiTaskSequential(taskDataSeq, "-");
+    ASSERT_EQ(testMpiTaskSequential.validation(), true);
+    testMpiTaskSequential.pre_processing();
+    testMpiTaskSequential.run();
+    testMpiTaskSequential.post_processing();
+
+    ASSERT_EQ(check_vec[0], global_res[0]);
+  }
+}
+
+TEST(varfolomeev_g_transfer_from_one_to_all_custom_scatter_mpi, Test_CustomScatter_Max_Zero_vec) {
+  boost::mpi::communicator world;
+  std::vector<int> global_vec;            // working vector
+  std::vector<int32_t> global_res(1, 0);  // result vector
+  std::shared_ptr<ppc::core::TaskData> taskDataPar = std::make_shared<ppc::core::TaskData>();
+
+  if (world.rank() == 0) {
+    const int count_size_vector = 50;
+    global_vec = varfolomeev_g_transfer_from_one_to_all_scatter_mpi::getRandomVector(count_size_vector, 0, 0);
+    taskDataPar->inputs.emplace_back(reinterpret_cast<uint8_t*>(global_vec.data()));
+    taskDataPar->inputs_count.emplace_back(global_vec.size());
+    taskDataPar->outputs.emplace_back(reinterpret_cast<uint8_t*>(global_res.data()));
+    taskDataPar->outputs_count.emplace_back(global_res.size());
+  }
+
+  varfolomeev_g_transfer_from_one_to_all_custom_scatter_mpi::MyScatterTestMPITaskParallel MyScatterTestMPITaskParallel(
+      taskDataPar, "max");
+  ASSERT_EQ(MyScatterTestMPITaskParallel.validation(), true);
+  MyScatterTestMPITaskParallel.pre_processing();
+  MyScatterTestMPITaskParallel.run();
+  MyScatterTestMPITaskParallel.post_processing();
+  if (world.rank() == 0) {
+    std::vector<int32_t> check_vec(1, 0);  // vector to check algorithm by seq. realisation
     std::shared_ptr<ppc::core::TaskData> taskDataSeq = std::make_shared<ppc::core::TaskData>();
     taskDataSeq->inputs.emplace_back(reinterpret_cast<uint8_t*>(global_vec.data()));
     taskDataSeq->inputs_count.emplace_back(global_vec.size());
@@ -232,7 +204,7 @@ TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Max_Zero_vec) {
   }
 }
 
-TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Sum_Negative) {
+TEST(varfolomeev_g_transfer_from_one_to_all_custom_scatter_mpi, Test_CustomScatter_Sum_Negative) {
   boost::mpi::communicator world;
   std::vector<int> global_vec;            // working vector
   std::vector<int32_t> global_res(1, 0);  // result vector
@@ -248,11 +220,12 @@ TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Sum_Negative) {
     taskDataPar->outputs_count.emplace_back(global_res.size());
   }
 
-  varfolomeev_g_transfer_from_one_to_all_scatter_mpi::TestMPITaskParallel testMpiTaskParallel(taskDataPar, "+");
-  ASSERT_EQ(testMpiTaskParallel.validation(), true);
-  testMpiTaskParallel.pre_processing();
-  testMpiTaskParallel.run();
-  testMpiTaskParallel.post_processing();
+  varfolomeev_g_transfer_from_one_to_all_custom_scatter_mpi::MyScatterTestMPITaskParallel MyScatterTestMPITaskParallel(
+      taskDataPar, "+");
+  ASSERT_EQ(MyScatterTestMPITaskParallel.validation(), true);
+  MyScatterTestMPITaskParallel.pre_processing();
+  MyScatterTestMPITaskParallel.run();
+  MyScatterTestMPITaskParallel.post_processing();
 
   if (world.rank() == 0) {
     std::vector<int32_t> check_vec(1, 0);  // vector to check algorithm by seq. realisation
@@ -262,17 +235,17 @@ TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Sum_Negative) {
     taskDataSeq->inputs_count.emplace_back(global_vec.size());
     taskDataSeq->outputs.emplace_back(reinterpret_cast<uint8_t*>(check_vec.data()));
     taskDataSeq->outputs_count.emplace_back(check_vec.size());
-    varfolomeev_g_transfer_from_one_to_all_scatter_mpi::TestMPITaskSequential testMpiTaskSequential(taskDataSeq, "+");
-    ASSERT_EQ(testMpiTaskSequential.validation(), true);
-    testMpiTaskSequential.pre_processing();
-    testMpiTaskSequential.run();
-    testMpiTaskSequential.post_processing();
+    varfolomeev_g_transfer_from_one_to_all_scatter_mpi::TestMPITaskSequential TestMPITaskSequential(taskDataSeq, "+");
+    ASSERT_EQ(TestMPITaskSequential.validation(), true);
+    TestMPITaskSequential.pre_processing();
+    TestMPITaskSequential.run();
+    TestMPITaskSequential.post_processing();
 
     ASSERT_EQ(check_vec[0], global_res[0]);
   }
 }
 
-TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Sum_Positive) {
+TEST(varfolomeev_g_transfer_from_one_to_all_custom_scatter_mpi, Test_CustomScatter_Sum_Positive) {
   boost::mpi::communicator world;
   std::vector<int> global_vec;            // working vector
   std::vector<int32_t> global_res(1, 0);  // result vector
@@ -288,11 +261,12 @@ TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Sum_Positive) {
     taskDataPar->outputs_count.emplace_back(global_res.size());
   }
 
-  varfolomeev_g_transfer_from_one_to_all_scatter_mpi::TestMPITaskParallel testMpiTaskParallel(taskDataPar, "+");
-  ASSERT_EQ(testMpiTaskParallel.validation(), true);
-  testMpiTaskParallel.pre_processing();
-  testMpiTaskParallel.run();
-  testMpiTaskParallel.post_processing();
+  varfolomeev_g_transfer_from_one_to_all_custom_scatter_mpi::MyScatterTestMPITaskParallel MyScatterTestMPITaskParallel(
+      taskDataPar, "+");
+  ASSERT_EQ(MyScatterTestMPITaskParallel.validation(), true);
+  MyScatterTestMPITaskParallel.pre_processing();
+  MyScatterTestMPITaskParallel.run();
+  MyScatterTestMPITaskParallel.post_processing();
 
   if (world.rank() == 0) {
     std::vector<int32_t> check_vec(1, 0);  // vector to check algorithm by seq. realisation
@@ -302,17 +276,17 @@ TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Sum_Positive) {
     taskDataSeq->inputs_count.emplace_back(global_vec.size());
     taskDataSeq->outputs.emplace_back(reinterpret_cast<uint8_t*>(check_vec.data()));
     taskDataSeq->outputs_count.emplace_back(check_vec.size());
-    varfolomeev_g_transfer_from_one_to_all_scatter_mpi::TestMPITaskSequential testMpiTaskSequential(taskDataSeq, "+");
-    ASSERT_EQ(testMpiTaskSequential.validation(), true);
-    testMpiTaskSequential.pre_processing();
-    testMpiTaskSequential.run();
-    testMpiTaskSequential.post_processing();
+    varfolomeev_g_transfer_from_one_to_all_scatter_mpi::TestMPITaskSequential TestMPITaskSequential(taskDataSeq, "+");
+    ASSERT_EQ(TestMPITaskSequential.validation(), true);
+    TestMPITaskSequential.pre_processing();
+    TestMPITaskSequential.run();
+    TestMPITaskSequential.post_processing();
 
     ASSERT_EQ(check_vec[0], global_res[0]);
   }
 }
 
-TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Sum_Positive_Negative) {
+TEST(varfolomeev_g_transfer_from_one_to_all_custom_scatter_mpi, Test_Sum_CustomScatter_Positive_Negative) {
   boost::mpi::communicator world;
   std::vector<int> global_vec;            // working vector
   std::vector<int32_t> global_res(1, 0);  // result vector
@@ -328,11 +302,12 @@ TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Sum_Positive_Negat
     taskDataPar->outputs_count.emplace_back(global_res.size());
   }
 
-  varfolomeev_g_transfer_from_one_to_all_scatter_mpi::TestMPITaskParallel testMpiTaskParallel(taskDataPar, "+");
-  ASSERT_EQ(testMpiTaskParallel.validation(), true);
-  testMpiTaskParallel.pre_processing();
-  testMpiTaskParallel.run();
-  testMpiTaskParallel.post_processing();
+  varfolomeev_g_transfer_from_one_to_all_custom_scatter_mpi::MyScatterTestMPITaskParallel MyScatterTestMPITaskParallel(taskDataPar,
+                                                                                                       "+");
+  ASSERT_EQ(MyScatterTestMPITaskParallel.validation(), true);
+  MyScatterTestMPITaskParallel.pre_processing();
+  MyScatterTestMPITaskParallel.run();
+  MyScatterTestMPITaskParallel.post_processing();
 
   if (world.rank() == 0) {
     std::vector<int32_t> check_vec(1, 0);  // vector to check algorithm by seq. realisation
@@ -352,7 +327,7 @@ TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Sum_Positive_Negat
   }
 }
 
-TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Diff_Negative) {
+TEST(varfolomeev_g_transfer_from_one_to_all_custom_scatter_mpi, Test_CustomScatter_Diff_Negative) {
   boost::mpi::communicator world;
   std::vector<int> global_vec;            // working vector
   std::vector<int32_t> global_res(1, 0);  // result vector
@@ -368,11 +343,12 @@ TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Diff_Negative) {
     taskDataPar->outputs_count.emplace_back(global_res.size());
   }
 
-  varfolomeev_g_transfer_from_one_to_all_scatter_mpi::TestMPITaskParallel testMpiTaskParallel(taskDataPar, "-");
-  ASSERT_EQ(testMpiTaskParallel.validation(), true);
-  testMpiTaskParallel.pre_processing();
-  testMpiTaskParallel.run();
-  testMpiTaskParallel.post_processing();
+  varfolomeev_g_transfer_from_one_to_all_custom_scatter_mpi::MyScatterTestMPITaskParallel MyScatterTestMPITaskParallel(
+      taskDataPar, "-");
+  ASSERT_EQ(MyScatterTestMPITaskParallel.validation(), true);
+  MyScatterTestMPITaskParallel.pre_processing();
+  MyScatterTestMPITaskParallel.run();
+  MyScatterTestMPITaskParallel.post_processing();
 
   if (world.rank() == 0) {
     std::vector<int32_t> check_vec(1, 0);  // vector to check algorithm by seq. realisation
@@ -392,7 +368,7 @@ TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Diff_Negative) {
   }
 }
 
-TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Diff_Positive) {
+TEST(varfolomeev_g_transfer_from_one_to_all_custom_scatter_mpi, Test_CustomScatter_Diff_Positive) {
   boost::mpi::communicator world;
   std::vector<int> global_vec;            // working vector
   std::vector<int32_t> global_res(1, 0);  // result vector
@@ -408,11 +384,12 @@ TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Diff_Positive) {
     taskDataPar->outputs_count.emplace_back(global_res.size());
   }
 
-  varfolomeev_g_transfer_from_one_to_all_scatter_mpi::TestMPITaskParallel testMpiTaskParallel(taskDataPar, "-");
-  ASSERT_EQ(testMpiTaskParallel.validation(), true);
-  testMpiTaskParallel.pre_processing();
-  testMpiTaskParallel.run();
-  testMpiTaskParallel.post_processing();
+  varfolomeev_g_transfer_from_one_to_all_custom_scatter_mpi::MyScatterTestMPITaskParallel MyScatterTestMPITaskParallel(
+      taskDataPar, "-");
+  ASSERT_EQ(MyScatterTestMPITaskParallel.validation(), true);
+  MyScatterTestMPITaskParallel.pre_processing();
+  MyScatterTestMPITaskParallel.run();
+  MyScatterTestMPITaskParallel.post_processing();
 
   if (world.rank() == 0) {
     std::vector<int32_t> check_vec(1, 0);  // vector to check algorithm by seq. realisation
@@ -432,7 +409,7 @@ TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Diff_Positive) {
   }
 }
 
-TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Diff_Positive_Negative) {
+TEST(varfolomeev_g_transfer_from_one_to_all_custom_scatter_mpi, Test_CustomScatter_Diff_Positive_Negative) {
   boost::mpi::communicator world;
   std::vector<int> global_vec;            // working vector
   std::vector<int32_t> global_res(1, 0);  // result vector
@@ -448,11 +425,12 @@ TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Diff_Positive_Nega
     taskDataPar->outputs_count.emplace_back(global_res.size());
   }
 
-  varfolomeev_g_transfer_from_one_to_all_scatter_mpi::TestMPITaskParallel testMpiTaskParallel(taskDataPar, "-");
-  ASSERT_EQ(testMpiTaskParallel.validation(), true);
-  testMpiTaskParallel.pre_processing();
-  testMpiTaskParallel.run();
-  testMpiTaskParallel.post_processing();
+  varfolomeev_g_transfer_from_one_to_all_custom_scatter_mpi::MyScatterTestMPITaskParallel MyScatterTestMPITaskParallel(
+      taskDataPar, "-");
+  ASSERT_EQ(MyScatterTestMPITaskParallel.validation(), true);
+  MyScatterTestMPITaskParallel.pre_processing();
+  MyScatterTestMPITaskParallel.run();
+  MyScatterTestMPITaskParallel.post_processing();
 
   if (world.rank() == 0) {
     std::vector<int32_t> check_vec(1, 0);  // vector to check algorithm by seq. realisation
@@ -472,7 +450,7 @@ TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Diff_Positive_Nega
   }
 }
 
-TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Max_Negative) {
+TEST(varfolomeev_g_transfer_from_one_to_all_custom_scatter_mpi, Test_CustomScatter_Max_Negative) {
   boost::mpi::communicator world;
   std::vector<int> global_vec;            // working vector
   std::vector<int32_t> global_res(1, 0);  // result vector
@@ -488,11 +466,12 @@ TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Max_Negative) {
     taskDataPar->outputs_count.emplace_back(global_res.size());
   }
 
-  varfolomeev_g_transfer_from_one_to_all_scatter_mpi::TestMPITaskParallel testMpiTaskParallel(taskDataPar, "max");
-  ASSERT_EQ(testMpiTaskParallel.validation(), true);
-  testMpiTaskParallel.pre_processing();
-  testMpiTaskParallel.run();
-  testMpiTaskParallel.post_processing();
+  varfolomeev_g_transfer_from_one_to_all_custom_scatter_mpi::MyScatterTestMPITaskParallel MyScatterTestMPITaskParallel(
+      taskDataPar, "max");
+  ASSERT_EQ(MyScatterTestMPITaskParallel.validation(), true);
+  MyScatterTestMPITaskParallel.pre_processing();
+  MyScatterTestMPITaskParallel.run();
+  MyScatterTestMPITaskParallel.post_processing();
 
   if (world.rank() == 0) {
     std::vector<int32_t> check_vec(1, 0);  // vector to check algorithm by seq. realisation
@@ -512,7 +491,7 @@ TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Max_Negative) {
   }
 }
 
-TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Max_Positive) {
+TEST(varfolomeev_g_transfer_from_one_to_all_custom_scatter_mpi, Test_CustomScatter_Max_Positive) {
   boost::mpi::communicator world;
   std::vector<int> global_vec;            // working vector
   std::vector<int32_t> global_res(1, 0);  // result vector
@@ -528,11 +507,12 @@ TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Max_Positive) {
     taskDataPar->outputs_count.emplace_back(global_res.size());
   }
 
-  varfolomeev_g_transfer_from_one_to_all_scatter_mpi::TestMPITaskParallel testMpiTaskParallel(taskDataPar, "max");
-  ASSERT_EQ(testMpiTaskParallel.validation(), true);
-  testMpiTaskParallel.pre_processing();
-  testMpiTaskParallel.run();
-  testMpiTaskParallel.post_processing();
+  varfolomeev_g_transfer_from_one_to_all_custom_scatter_mpi::MyScatterTestMPITaskParallel MyScatterTestMPITaskParallel(
+      taskDataPar, "max");
+  ASSERT_EQ(MyScatterTestMPITaskParallel.validation(), true);
+  MyScatterTestMPITaskParallel.pre_processing();
+  MyScatterTestMPITaskParallel.run();
+  MyScatterTestMPITaskParallel.post_processing();
 
   if (world.rank() == 0) {
     std::vector<int32_t> check_vec(1, 0);  // vector to check algorithm by seq. realisation
@@ -552,7 +532,7 @@ TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Max_Positive) {
   }
 }
 
-TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Max_Positive_Negative) {
+TEST(varfolomeev_g_transfer_from_one_to_all_custom_scatter_mpi, Test_CustomScatter_Max_Positive_Negative) {
   boost::mpi::communicator world;
   std::vector<int> global_vec;            // working vector
   std::vector<int32_t> global_res(1, 0);  // result vector
@@ -568,11 +548,12 @@ TEST(varfolomeev_g_transfer_from_one_to_all_scatter_mpi, Test_Max_Positive_Negat
     taskDataPar->outputs_count.emplace_back(global_res.size());
   }
 
-  varfolomeev_g_transfer_from_one_to_all_scatter_mpi::TestMPITaskParallel testMpiTaskParallel(taskDataPar, "max");
-  ASSERT_EQ(testMpiTaskParallel.validation(), true);
-  testMpiTaskParallel.pre_processing();
-  testMpiTaskParallel.run();
-  testMpiTaskParallel.post_processing();
+  varfolomeev_g_transfer_from_one_to_all_custom_scatter_mpi::MyScatterTestMPITaskParallel MyScatterTestMPITaskParallel(
+      taskDataPar, "max");
+  ASSERT_EQ(MyScatterTestMPITaskParallel.validation(), true);
+  MyScatterTestMPITaskParallel.pre_processing();
+  MyScatterTestMPITaskParallel.run();
+  MyScatterTestMPITaskParallel.post_processing();
 
   if (world.rank() == 0) {
     std::vector<int32_t> check_vec(1, 0);  // vector to check algorithm by seq. realisation
